third_party/WebKit/LayoutTests/svg/animations/reinserting-svg-into-document-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
Reinserting SVG animation into document should continue the animation

PASS successfullyParsed is true

TEST COMPLETE
PASS rect.x.animVal.value is 30
PASS rect.x.animVal.value is 60
